fixed image attachments and others
Fixed <img> tags not including the image in the downloaded epub file

Added replacing <img> with <img /> to make xhtml happy

Fixed end-of-chapter detection incorrectly detecting any double <hr> as
the end of the chapter. Now the double <hr> must be directly followed by
the words "end" and "chapter" for the detection to kick in.